requirements.txt
zelos-demeter==0.7.5
trading-strategy==0.24.4
scipy==1.14.1